refactor: move util test functions to separate file
