+++ core/Cargo.toml
@@ -26,4 +26,3 @@ tokio = { version = "1.21.2", features = ["macros"] }
 serde_json = "1.0"
 serde = { version = "1.0" }
 wiremock = "0.5.17"
-serde_with = "2.2.0"
