+++ core/src/utils/wiremock_utils.rs
@@ -1,21 +1,13 @@
 use std::str::FromStr;
 
+use crate::helpers::ethers_block_id_to_starknet_block_id;
 use reth_primitives::rpc::{BlockId, H256};
-use serde::{de::Error as DeError, Deserialize, Deserializer, Serialize, Serializer};
-use serde_with::{serde_as, DeserializeAs, SerializeAs};
-use starknet::{
-    core::types::FieldElement,
-    providers::jsonrpc::{
-        models::{BlockId as StarknetBlockId, BlockTag},
-        JsonRpcMethod,
-    },
-};
+use serde::{Deserialize, Serialize};
+use starknet::providers::jsonrpc::models::{BlockId as StarknetBlockId, BlockTag};
 use wiremock::{
     matchers::{body_json, method},
     Mock, MockServer, ResponseTemplate,
 };
-
-use crate::helpers::ethers_block_id_to_starknet_block_id;
 
 #[derive(Serialize, Debug)]
 pub struct StarknetRpcBaseData<'a, StarknetParams> {
@@ -106,59 +98,6 @@ impl<'a, StarknetParams> StarknetRpcBaseData<'a, StarknetParams> {
         }
     }
 }
-
-// DEBUG FUNCTION
-pub struct UfeHex;
-
-impl SerializeAs<FieldElement> for UfeHex {
-    fn serialize_as<S>(value: &FieldElement, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        serializer.serialize_str(&format!("{value:#x}"))
-    }
-}
-
-impl<'de> DeserializeAs<'de, FieldElement> for UfeHex {
-    fn deserialize_as<D>(deserializer: D) -> Result<FieldElement, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let value = String::deserialize(deserializer)?;
-        match FieldElement::from_hex_be(&value) {
-            Ok(value) => Ok(value),
-            Err(err) => Err(DeError::custom(format!("invalid hex string: {err}"))),
-        }
-    }
-}
-
-#[serde_as]
-#[derive(Serialize, Deserialize)]
-struct Felt(#[serde_as(as = "UfeHex")] pub FieldElement);
-
-#[derive(Debug, Serialize)]
-struct JsonRpcRequest<T> {
-    id: u64,
-    jsonrpc: &'static str,
-    method: JsonRpcMethod,
-    params: T,
-}
-
-pub fn debug_entry_param(method: JsonRpcMethod, params: Vec<serde_json::Value>) -> String {
-    let request = JsonRpcRequest {
-        id: 1,
-        jsonrpc: "2.0",
-        method,
-        params,
-    };
-    println!(
-        "JsonRpcRequest: {:?}",
-        serde_json::to_string(&request).unwrap()
-    );
-    serde_json::to_string(&request).unwrap()
-}
-
-// END DEBUG FUNCTION
 
 pub async fn setup_wiremock() -> String {
     let mock_server = MockServer::start().await;
