+++ rpc/Cargo.toml
@@ -35,6 +35,7 @@ reqwest = "0.11.13"
 dotenv = "0.15.0"
 mockall = "0.11.3"
 anyhow = "1.0.68"
+serde_with = "2.2.0"
 
 [dev-dependencies.cargo-husky]
 version = "1.5.0"
